converted /pages/api to typescript
