+++ pages/api/residents/[id]/allocations/[allocationId].ts
@@ -3,7 +3,12 @@ import { StatusCodes } from 'http-status-codes';
 import { getResidentAllocation } from 'utils/server/allocatedWorkers';
 import { isAuthorised } from 'utils/auth';
 
-export default async (req, res) => {
+import type { NextApiRequest, NextApiResponse, NextApiHandler } from 'next';
+
+const endpoint: NextApiHandler = async (
+  req: NextApiRequest,
+  res: NextApiResponse
+) => {
   const user = isAuthorised(req);
   if (!user) {
     return res.status(StatusCodes.UNAUTHORIZED).end();
@@ -38,3 +43,5 @@ export default async (req, res) => {
       console.error(res.status);
   }
 };
+
+export default endpoint;
