+++ pages/api/residents/[id]/allocations/index.ts
@@ -7,7 +7,12 @@ import {
 } from 'utils/server/allocatedWorkers';
 import { isAuthorised } from 'utils/auth';
 
-export default async (req, res) => {
+import type { NextApiRequest, NextApiResponse, NextApiHandler } from 'next';
+
+const endpoint: NextApiHandler = async (
+  req: NextApiRequest,
+  res: NextApiResponse
+) => {
   const user = isAuthorised(req);
   if (!user) {
     return res.status(StatusCodes.UNAUTHORIZED).end();
@@ -81,3 +86,5 @@ export default async (req, res) => {
       console.error(res.status);
   }
 };
+
+export default endpoint;
