+++ pages/api/residents/[id]/cases/index.ts
@@ -1,9 +1,14 @@
 import { StatusCodes } from 'http-status-codes';
 
+import { getCasesByResident } from 'utils/server/cases';
 import { isAuthorised } from 'utils/auth';
-import { getAddresses } from 'utils/server/postcode';
 
-export default async (req, res) => {
+import type { NextApiRequest, NextApiResponse, NextApiHandler } from 'next';
+
+const endpoint: NextApiHandler = async (
+  req: NextApiRequest,
+  res: NextApiResponse
+) => {
   const user = isAuthorised(req);
   if (!user) {
     return res.status(StatusCodes.UNAUTHORIZED).end();
@@ -11,16 +16,20 @@ export default async (req, res) => {
   if (!user.isAuthorised) {
     return res.status(StatusCodes.FORBIDDEN).end();
   }
+  const { id, ...params } = req.query;
   switch (req.method) {
     case 'GET':
       try {
-        const data = await getAddresses(req.query.postcode);
+        const data = await getCasesByResident(id, {
+          ...params,
+          context_flag: user.permissionFlag,
+        });
         res.status(StatusCodes.OK).json(data);
       } catch (error) {
-        console.error('Postcode get error', error?.response?.data);
+        console.error('Cases get error:', error?.response?.data);
         res
           .status(StatusCodes.INTERNAL_SERVER_ERROR)
-          .json({ message: 'Unable to get the Addresses' });
+          .json({ message: 'Unable to get the Cases' });
       }
       break;
 
@@ -30,3 +39,5 @@ export default async (req, res) => {
         .json({ message: 'Invalid request method' });
   }
 };
+
+export default endpoint;
