+++ pages/api/residents/index.ts
@@ -1,0 +1,58 @@
+import { StatusCodes } from 'http-status-codes';
+
+import { getResidents, addResident } from 'utils/server/residents';
+import { isAuthorised } from 'utils/auth';
+
+import type { NextApiRequest, NextApiResponse, NextApiHandler } from 'next';
+
+const endpoint: NextApiHandler = async (
+  req: NextApiRequest,
+  res: NextApiResponse
+) => {
+  const user = isAuthorised(req);
+  if (!user) {
+    return res.status(StatusCodes.UNAUTHORIZED).end();
+  }
+  if (!user.isAuthorised) {
+    return res.status(StatusCodes.FORBIDDEN).end();
+  }
+  switch (req.method) {
+    case 'GET':
+      try {
+        const data = await getResidents({
+          ...req.query,
+          context_flag: user.permissionFlag,
+        });
+        res.status(StatusCodes.OK).json(data);
+      } catch (error) {
+        console.error('Residents get error:', error?.response?.data);
+        error?.response?.status === StatusCodes.NOT_FOUND
+          ? res
+              .status(StatusCodes.NOT_FOUND)
+              .json({ message: 'Residents Not Found' })
+          : res
+              .status(StatusCodes.INTERNAL_SERVER_ERROR)
+              .json({ message: 'Unable to get the Residents' });
+      }
+      break;
+
+    case 'POST':
+      try {
+        const data = await addResident(req.body);
+        res.status(StatusCodes.OK).json(data);
+      } catch (error) {
+        console.error('Resident post error:', error?.response?.data);
+        res
+          .status(StatusCodes.INTERNAL_SERVER_ERROR)
+          .json({ message: 'Unable to add resident' });
+      }
+      break;
+
+    default:
+      res
+        .status(StatusCodes.BAD_REQUEST)
+        .json({ message: 'Invalid request method' });
+  }
+};
+
+export default endpoint;
