+++ pages/api/teams/[team_id]/workers.ts
@@ -1,9 +1,14 @@
 import { StatusCodes } from 'http-status-codes';
 
-import { getCases, addCase } from 'utils/server/cases';
+import { getWorkers } from 'utils/server/workers';
 import { isAuthorised } from 'utils/auth';
 
-export default async (req, res) => {
+import type { NextApiRequest, NextApiResponse, NextApiHandler } from 'next';
+
+const endpoint: NextApiHandler = async (
+  req: NextApiRequest,
+  res: NextApiResponse
+) => {
   const user = isAuthorised(req);
   if (!user) {
     return res.status(StatusCodes.UNAUTHORIZED).end();
@@ -14,32 +19,17 @@ export default async (req, res) => {
   switch (req.method) {
     case 'GET':
       try {
-        const data = await getCases({
-          ...req.query,
-          context_flag: user.permissionFlag,
-        });
+        const data = await getWorkers(req.query);
         res.status(StatusCodes.OK).json(data);
       } catch (error) {
-        console.error('Cases get error:', error?.response?.data);
+        console.error('Workers get error:', error?.response?.data);
         error?.response?.status === StatusCodes.NOT_FOUND
           ? res
               .status(StatusCodes.NOT_FOUND)
-              .json({ message: 'Cases Not Found' })
+              .json({ message: 'Workers Not Found' })
           : res
               .status(StatusCodes.INTERNAL_SERVER_ERROR)
-              .json({ message: 'Unable to get the Cases' });
-      }
-      break;
-
-    case 'POST':
-      try {
-        const data = await addCase(req.body);
-        res.status(StatusCodes.OK).json(data);
-      } catch (error) {
-        console.error('Case post error:', error?.response?.data);
-        res
-          .status(StatusCodes.INTERNAL_SERVER_ERROR)
-          .json({ message: 'Unable to post case' });
+              .json({ message: 'Unable to get the Workers' });
       }
       break;
 
@@ -49,3 +39,5 @@ export default async (req, res) => {
         .json({ message: 'Invalid request method' });
   }
 };
+
+export default endpoint;
